backend/requirement.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
Faker==14.2.0
numpy==1.23.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
